pyproject.toml
[project]
name = "geotoolkit"
version = "0.1.0"
description = "Automated Security Scanner for Git Repositories"
readme = "README.md"
authors = []

requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "pydantic>=2.10.6",
    "swagger-ui>=0.1.2",
    "uvicorn[standard]>=0.34.0",
    "waitress>=3.0.2",
    "Jinja2>=3.1.4",
    "requests>=2.32.3",
    "gitpython>=3.1.0",
]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Development Status :: 3 - Alpha"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[project.scripts]
geotoolkit = "src.main:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
    "mypy>=1.6.0",
    "pytest>=7.3.1",
    "hatchling",
    "types-waitress"
]
test = [
    "ruff>=0.11.0",
    "mypy>=1.6.0",
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "requests>=2.32.3",
]

[tool.hatch.build]
artifact-name = "geotoolkit-{version}"
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "src/reporting/templates/report.md",
    "tests",
]

[tool.hatch.envs.default]
dependencies = [
    "ruff>=0.11.0",
    "mypy>=1.6.0",
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "types-waitress"
]
[tool.hatch.envs.default.scripts]
test = "pytest --cov=src tests"
lint = "ruff check src tests pyproject.toml && mypy src tests"
format = "ruff format src tests pyproject.toml"
run = "uvicorn src.main:app --reload"
clean = "rm -rf .coverage .pytest_cache dist build"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B", "Q", "UP", "YTT", "PYI"]
ignore = ["E501", "B008", "B905"] 


[tool.ruff]
target-version = "py311" 
src = ["src", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["error"]

[tool.mypy]
python_version = "3.11"
show_error_codes = true
pretty = true
strict = true

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["src/reporting/templates/*"] 

[tool.coverage.paths]
source = ["src"]
